feat(web): step-by-step /setup wizard + Built-in AI settings page
Splits the agent-setup UX into two surfaces with different chrome:

- /setup → AgentSetupWizard. Three-step flow with stepper:
    Step 1 Provider — preset, URL, API key + Save & test
    Step 2 Models   — chat + curator pickers
    Step 3 Lift-off — checklist + CTAs (Go to inbox, Tweak settings)
  "Wake Munin." in the hero, italicized accent matching the inbox
  treatment. Smart entry: provider key set → jump to Step 3,
  otherwise Step 1. Stepper dots let you walk back.

- /dashboard/settings/ai-agent → AgentSettingsPage. Single-form
  layout, plain "Built-in AI" hero, no stepper. Lives inside the
  Settings rail layout (no own <main> wrapper, layout owns padding).

Shared primitives in packages/dashboard-pages/src/components/agent-config/:

- types.ts        — AgentConfigDto, ModelEntry, PROVIDER_PRESETS,
                    presetForUrl, formatModel
- provider-card.tsx
- models-card.tsx — accepts saveLabel + extraActions for the
                    wizard's Save & continue + Back
- use-agent-config.ts — load + auto-fetch /v1/models when key set

Settings nav entry renamed "AI agent" → "Built-in AI" (Innebygd AI
in nb) under Workspace, between Channels and Data export. The "AI
agent" label collided with the existing "Connected agents" entry
two rows below; the new label disambiguates and covers both
chat-reply and curation responsibilities.

Drops AgentSetupPage in favor of AgentSettingsPage + AgentSetupWizard.
