+++ apps/web/app/dashboard/settings/ai-agent/page.tsx
@@ -1,0 +1,1 @@
+export { AgentSettingsPage as default } from '@getmunin/dashboard-pages';
