+++ packages/dashboard-pages/src/index.ts
@@ -5,7 +5,8 @@ export { PageShell, nativeFieldClass } from './components/page-shell';
 export { useActiveRole, isOwnerOrAdmin, type OrgRole } from './auth/use-active-role';
 
 export { AcceptInvitePage } from './pages/accept-invite';
-export { AgentSetupPage } from './pages/agent-setup';
+export { AgentSettingsPage } from './pages/agent-settings';
+export { AgentSetupWizard } from './pages/agent-setup-wizard';
 export { AgentsPage } from './pages/agents';
 export { ApiKeysPage } from './pages/api-keys';
 export { AuditLogPage } from './pages/audit-log';
